move new function to gen_utils, its cleaner
